Add flatpak files from Audiveris-flatpak
Copy the files and helper scripts from
https://github.com/mwilck/Audiveris-flatpak into the dev/flatpak
subdirectory, omitting files that are duplicates of files in the
dev/flathub submodule.

Compared with the code that was used to create the Audiveris 5.3.1 flathub
submission, the deps.py script has been extended to detect x86_64 specific
artifacts, and automatically add the arm64 equivalents of these artifacts
to the artifact list. This may allow us to remove the restriction that
audiveris is only built for x86_64 in the future.

Adapt .gitignore, too.