Flatpak: remove filesystem permissions for xdg-config etc.
Access to xdg-cache etc. without subdirectory is not allowed by the flatpak
linter.

https://docs.flathub.org/docs/for-app-authors/linter/#finish-args-arbitrary-xdg-cache-mode-access
https://docs.flathub.org/docs/for-app-authors/linter/#finish-args-arbitrary-xdg-config-mode-access
https://docs.flathub.org/docs/for-app-authors/linter/#finish-args-arbitrary-xdg-data-mode-access

I tried to change this to "--filesystem=xdg-config/AudiverisLtd:create" etc.,
but the linter doesn't accept that as well, even though applying for an
exception would be possible with that setup.

https://docs.flathub.org/docs/for-app-authors/linter/#finish-args-unnecessary-xdg-cache-subdir-mode-access
https://docs.flathub.org/docs/for-app-authors/linter/#finish-args-unnecessary-xdg-config-subdir-mode-access
https://docs.flathub.org/docs/for-app-authors/linter/#finish-args-unnecessary-xdg-data-subdir-mode-access

Therefore, remove these permissions for now, they aren't strictly necessary.